Introducing multi-staged builds to replace manual build of Go app
