--- Dockerfile.alpine
@@ -1,5 +1,0 @@
-FROM alpine
-
-ADD mysql_health_check /
-
-CMD ["/mysql_health_check"]
