+++ Dockerfile.scratch
@@ -1,5 +1,19 @@
+FROM golang:alpine3.7 as builder
+
+LABEL maintainer='Hugo Fonseca <https://github.com/hugomcfonseca>'
+
+WORKDIR /go/src/github.com/hugomcfonseca/mysql_health_api/
+
+COPY /app/mysql_health_check.go .
+
+RUN apk add --update --no-cache git \
+    && go get -d -v \
+    && CGO_ENABLED=0 GOOS=linux go build --ldflags '-s' -a -installsuffix cgo -o mysql_health_check .
+
 FROM scratch
 
-ADD mysql_health_check /
+LABEL maintainer='Hugo Fonseca <https://github.com/hugomcfonseca>'
 
-CMD ["/mysql_health_check"]
+COPY --from=builder /go/src/github.com/hugomcfonseca/mysql_health_api/mysql_health_check /
+
+ENTRYPOINT [ "mysql_health_check" ]
